// src/pages/link/data.d.ts
/*
 * +----------------------------------------------------------------------------------
 * | https://www.tigervs.com
 * +----------------------------------------------------------------------------------
 * | Email: Kevin@tigervs.com
 * +----------------------------------------------------------------------------------
 * | Copyright (c) Shenzhen Tiger Technology Co., Ltd. 2018~2022. All rights reserved.
 * +----------------------------------------------------------------------------------
 */

/** @format */

export type linkDataItem = {
    url?: string;
    sort?: number;
    name?: string;
    status?: number;
    contact?: number;
    id?: number | string;
    description?: string;
    create_time?: string;
    update_time?: string;
};